repo: reorganize directory structure
reorganize directory structure
