+++ pkg/k8sutil/configmap.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package kubernetes
+package k8sutil
 
 import (
 	"context"
